change figured outgit add -A

+++ capstone/src/main/java/com/techelevator/view/Dessert.java
@@ -1,7 +1,9 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Dessert extends Food {
-    public Dessert(String name, double price) {
+    public Dessert(String name, BigDecimal price) {
         super(name, price);
     }
